requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.0
huey==2.4.5
redis==5.0.1
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
